add basic api for graph implementation, disable travis on kg_graph branch
