+++ etk/knowledge_graph/ontology.py
@@ -1,0 +1,5 @@
+from etk.knowledge_graph.graph import Graph
+
+
+class Ontology(Graph):
+    pass
