+++ etk/knowledge_graph/ontology_namespacemanager.py
